requirements.txt
# Experimental discord.py with UI Kit Components v2 from pull request
git+https://github.com/Rapptz/discord.py.git@feature/voice-receive#egg=discord.py

# Alternative stable requirements (when v2 components are officially released):
# discord.py>=2.6.0

# Other dependencies
requests>=2.25.1
python-dotenv>=0.19.0
